Simplify game field creation method header

+++ app/src/main/java/com/shpakovskiy/dynamicocean/model/GameField.kt
@@ -2,8 +2,8 @@ package com.shpakovskiy.dynamicocean.model
 
 // Common rect structure
 data class GameField(
-    val x: Float,
-    val y: Float,
+    val x: Int,
+    val y: Int,
     val heightCollapsed: Int,
     val widthCollapsed: Int,
     val heightExpanded: Int,
